Backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
httpx
pathlib
pydantic
python-multipart
requests